+++ src/dropin-model.js
@@ -132,9 +132,14 @@ DropinModel.prototype.changeActivePaymentMethod = function (paymentMethod) {
   this._emit('changeActivePaymentMethod', paymentMethod);
 };
 
-DropinModel.prototype.changeActivePaymentView = function (paymentViewID) {
-  this._activePaymentView = paymentViewID;
-  this._emit('changeActivePaymentView', paymentViewID);
+DropinModel.prototype.changeActiveView = function (paymentViewID) {
+  var previousViewId = this._activePaymentViewId;
+
+  this._activePaymentViewId = paymentViewID;
+  this._emit('changeActiveView', {
+    previousViewId: previousViewId,
+    newViewId: paymentViewID
+  });
 };
 
 DropinModel.prototype.removeActivePaymentMethod = function () {
@@ -229,8 +234,8 @@ DropinModel.prototype.getActivePaymentMethod = function () {
   return this._activePaymentMethod;
 };
 
-DropinModel.prototype.getActivePaymentView = function () {
-  return this._activePaymentView;
+DropinModel.prototype.getActivePaymentViewId = function () {
+  return this._activePaymentViewId;
 };
 
 DropinModel.prototype.reportAppSwitchPayload = function (payload) {